Initial scaffolding for creating GTP requests and "translate" those to Gatling Actions
Base classes for GTP request builders, which is then producing a Gatling
Action Builder, which in turn will generate an Action, which is what
Gatling will execute in order to have the actual message sent.

Currently, the Action doesn't do anything since this is just wiring it
all up but the pattern is there and you can actually run a scenario,
even though, again, it won't actually build the actual GTP message and
as such, won't send it across the wire.

Next is to actually start the GTP Stack, wire it up through the builders
so that the GtpAction can use the stack to actually send the message.